+++ src/app/admin/components/index.ts
@@ -1,0 +1,5 @@
+export * from './admin-dashboard/admin-dashboard.component';
+export * from './manage-products/manage-products.component';
+export * from './manage-orders/manage-orders.component';
+export * from './product-form/product-form.component';
+export * from './product-item/product-item.component';