+++ src/app/cart/models/cart-products.models.ts
@@ -1,9 +1,6 @@
-import { CartModel } from './cart.model';
-
 export class CartProductsModel {
 
     constructor(
-        public cartProducts: Array<CartModel>,
         public totalQuantity: number,
         public totalSum: number) { }
 }