+++ src/app/core/services/index.ts
@@ -3,3 +3,4 @@ export * from './constant.service';
 export * from './generator.factory';
 export * from './generator';
 export * from './local-storage.service';
+export * from './auth.service';