+++ src/app/layout/components/about/about.component.ts
@@ -1,5 +1,4 @@
 import { Component, Optional, Inject, OnInit } from '@angular/core';
-import { LocalStorageService } from '../../core/services/local-storage.service';
 import {
   ConfigOptionsService,
   ConstantsServiceToken,
@@ -7,7 +6,7 @@ import {
   GeneratorService,
   RandomString,
   GeneratorFactory
- } from '../../core';
+ } from '../../../core';
 
 const constantService = new ConstantsService('Shop', '1.0');
 
@@ -17,7 +16,6 @@ const constantService = new ConstantsService('Shop', '1.0');
   styleUrls: ['./about.component.css'],
   providers: [
     GeneratorService,
-    { provide: LocalStorageService, useClass: LocalStorageService },
     { provide: ConstantsServiceToken, useValue: constantService },
     { provide: RandomString, useFactory:  GeneratorFactory(3), deps: [GeneratorService] }
   ]
@@ -27,7 +25,6 @@ export class AboutComponent implements OnInit{
   version: string;
 
   constructor(
-    @Optional() private localStorageService: LocalStorageService,
     @Optional() private configOptionsService: ConfigOptionsService,
     @Inject(ConstantsServiceToken)@Optional() private constantsService: ConstantsService,
     @Inject(RandomString)@Optional() private randomString: string) {