+++ src/app/orders/orders.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+
+import { SharedModule } from '../shared/shared.module';
+import { OrdersRoutingModule } from './orders-routing.module';
+
+@NgModule({
+  declarations: [
+    OrdersRoutingModule.components
+  ],
+  imports: [
+    SharedModule,
+    OrdersRoutingModule
+  ]
+})
+export class OrdersModule { }